single-actor tests
